update: add function Add&Done, support graceful shutdown
