+++ tao_test.go
@@ -1,4 +1,4 @@
-// Copyright 2021 huija
+// Copyright 2022 huija
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -16,9 +16,37 @@ package tao
 
 import (
 	"context"
+	"encoding/json"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
+
+func TestRegister(t *testing.T) {
+	err := Register(printConfigKey, func() error {
+		p := new(printConfig)
+		// 1. transfer config bytes to object
+		bytes, err := GetConfigBytes(printConfigKey)
+		if err != nil {
+			return err
+		}
+		err = json.Unmarshal(bytes, &p)
+		if err != nil {
+			return err
+		}
+
+		p.ValidSelf()
+
+		// 2. set object to tao
+		return SetConfig(printConfigKey, p)
+	})
+	assert.Nil(t, err)
+
+	err = SetConfig(printConfigKey, nil)
+	assert.NotNil(t, err)
+
+	err = universeInit()
+	assert.NotNil(t, err)
+}
 
 func TestRun(t *testing.T) {
 	t.Log(new(taoConfig).Default())
@@ -30,6 +58,8 @@ func TestRun(t *testing.T) {
 	err := Run(cancel, nil)
 	assert.NotNil(t, err)
 
+	Add(1)
+	Done()
 	err = Run(nil, nil)
 	assert.Nil(t, err)
 }
